fixed issue #11
