+++ src/styles/find_poll.scss
@@ -1,0 +1,15 @@
+article {
+  section {
+    width: fit-content;
+    height: fit-content;
+    background-color: #f1fec6;
+    border-radius: 6px;
+    padding-right: 15px;
+    padding-left: 15px;
+    padding-bottom: 10px;
+    text-align: center;
+    input {
+      border: 1px #bfdf8e solid;
+    }
+  }
+}
